Added Milkor grenade launcher
  Since it's been awhile I figured let's add one of the deadly grenade
launchers.

  -- Great ideas follow --

  > P.S. I'm wondering if we could use the grenade ammo (40mm) in the
crafting recipe for the grenade launchers?

  > Also P.S. For weapons like the Golden Desert Eagle we could simply
require gold blocks to "upgrade" a regular Desert Eagle to this super
cool looking version. (This could apply say iron blocks to "upgrade" the
Desert Eagle into the Python too, kind of a alt upgrade options for the
tier 3s)

  > We could put some kind of resource cost onto say the Intervention
(m200) and have it "upgrade" to the G11?

  > And maybe a bunch of 40mm grenades placed with a M79 to "upgrade" to
the Milkor?